Moved knot.x core to knotx-core module, created knotx-example module deploying mock service verticle and knot.x core
