+++ knotx-core/pom.xml
@@ -23,11 +23,11 @@
     <modelVersion>4.0.0</modelVersion>
 	<parent>
 		<groupId>com.cognifide.knotx</groupId>
-		<artifactId>knot.x-root</artifactId>
+		<artifactId>knotx-root</artifactId>
 		<version>1.0.0-SNAPSHOT</version>
 	</parent>
 
-	<artifactId>knot.x</artifactId>
+	<artifactId>knotx-core</artifactId>
 	<version>1.0.0-SNAPSHOT</version>
 
 	<name>Knot.x - Reactive microservice assembler</name>
@@ -113,6 +113,9 @@
 			<plugin>
 				<groupId>org.springframework.boot</groupId>
 				<artifactId>spring-boot-maven-plugin</artifactId>
+				<configuration>
+					<skip>true</skip>
+				</configuration>
 			</plugin>
 			<!-- plugin for generating licensing header in headers of defined files -->
 			<plugin>
@@ -132,4 +135,21 @@
 			</plugin>
 		</plugins>
 	</build>
+
+	<profiles>
+		<profile>
+			<id>core</id>
+			<build>
+				<plugins>
+					<plugin>
+						<groupId>org.springframework.boot</groupId>
+						<artifactId>spring-boot-maven-plugin</artifactId>
+						<configuration>
+							<skip>false</skip>
+						</configuration>
+					</plugin>
+				</plugins>
+			</build>
+		</profile>
+	</profiles>
 </project>
