+++ knotx-core/src/main/java/com/cognifide/knotx/Application.java
