+++ knotx-core/src/main/java/com/cognifide/knotx/KnotxConfiguration.java
