+++ knotx-core/src/main/java/com/cognifide/knotx/KnotxVerticle.java
