+++ knotx-core/src/main/java/com/cognifide/knotx/event/TrafficObserver.java
