+++ knotx-core/src/main/java/com/cognifide/knotx/handlebars/Helpers.java
