+++ knotx-core/src/main/java/com/cognifide/knotx/handler/RestServiceResponseHandler.java
