+++ knotx-core/src/main/java/com/cognifide/knotx/repository/BasicTemplate.java
