+++ knotx-core/src/main/java/com/cognifide/knotx/repository/RepositoryConfiguration.java
