+++ knotx-core/src/main/java/com/cognifide/knotx/service/MockServiceHandler.java
