+++ knotx-core/src/main/java/com/cognifide/knotx/service/ServiceEndpointFacade.java
