+++ knotx-core/src/main/java/com/cognifide/knotx/template/TemplateHandlerFactory.java
