+++ knotx-core/src/main/resources/logback.xml
