+++ knotx-core/src/main/resources/repository.yml
