+++ knotx-core/src/main/resources/service.yml
