+++ knotx-example/src/main/resources/application.yml
@@ -1,5 +1,5 @@
 #
-# Knot.x - Mock service
+# Knot.x - Sample App with Mock service
 #
 # Copyright (C) 2016 Cognifide Limited
 #
@@ -39,3 +39,7 @@ service:
 
 repository:
   configuration: classpath:repository.yml
+
+spring:
+  main:
+    web-environment: false
