+++ knotx-example/src/main/resources/content/local/simple.html
@@ -1,6 +1,6 @@
 <!--
 
-    Knot.x - Mock service
+    Knot.x - Sample App with Mock service
 
     Copyright (C) 2016 Cognifide Limited
