+++ knotx-example/src/main/resources/logback.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
 
-    Knot.x - Mock service
+    Knot.x - Sample App with Mock service
 
     Copyright (C) 2016 Cognifide Limited
